+++ src/main/java/com/prgrms/mukvengers/domain/store/service/StoreService.java
@@ -1,11 +1,40 @@
 package com.prgrms.mukvengers.domain.store.service;
 
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import com.prgrms.mukvengers.domain.store.dto.request.CreateStoreRequest;
 import com.prgrms.mukvengers.domain.store.dto.response.StoreResponse;
+import com.prgrms.mukvengers.domain.store.exception.StoreNotFoundException;
+import com.prgrms.mukvengers.domain.store.mapper.StoreMapper;
+import com.prgrms.mukvengers.domain.store.model.Store;
+import com.prgrms.mukvengers.domain.store.repository.StoreRepository;
 
-public interface StoreService {
+import lombok.RequiredArgsConstructor;
 
-	String create(CreateStoreRequest createStoreRequest);
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class StoreService {
 
-	StoreResponse getId(Long placeId);
+	private final StoreRepository storeRepository;
+	private final StoreMapper storeMapper;
+
+	@Transactional
+	public Long create(CreateStoreRequest createStoreRequest) {
+
+		Store store = storeMapper.toStore(createStoreRequest);
+		storeRepository.save(store);
+
+		return store.getId();
+	}
+
+	public StoreResponse getById(Long storeId) {
+
+		Store store = storeRepository.findById(storeId)
+			.orElseThrow(() -> new StoreNotFoundException(storeId));
+
+		return storeMapper.toStoreResponse(store);
+	}
+
 }
