+++ src/test/java/com/prgrms/mukvengers/base/ServiceTest.java
@@ -3,6 +3,8 @@ package com.prgrms.mukvengers.base;
 import static com.prgrms.mukvengers.utils.StoreObjectProvider.*;
 import static com.prgrms.mukvengers.utils.UserObjectProvider.*;
 
+import com.prgrms.mukvengers.domain.crew.service.CrewService;
+import com.prgrms.mukvengers.domain.store.service.StoreService;
 import org.junit.jupiter.api.BeforeEach;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,7 +14,6 @@ import org.springframework.transaction.annotation.Transactional;
 import com.prgrms.mukvengers.domain.chat.repository.ChatRepository;
 import com.prgrms.mukvengers.domain.chat.service.ChatService;
 import com.prgrms.mukvengers.domain.crew.repository.CrewRepository;
-import com.prgrms.mukvengers.domain.crew.service.CrewService;
 import com.prgrms.mukvengers.domain.crewmember.repository.CrewMemberRepository;
 import com.prgrms.mukvengers.domain.proposal.repository.ProposalRepository;
 import com.prgrms.mukvengers.domain.proposal.service.ProposalService;
@@ -20,7 +21,6 @@ import com.prgrms.mukvengers.domain.review.repository.ReviewRepository;
 import com.prgrms.mukvengers.domain.review.service.ReviewService;
 import com.prgrms.mukvengers.domain.store.model.Store;
 import com.prgrms.mukvengers.domain.store.repository.StoreRepository;
-import com.prgrms.mukvengers.domain.store.service.StoreService;
 import com.prgrms.mukvengers.domain.user.model.User;
 import com.prgrms.mukvengers.domain.user.repository.UserRepository;
 import com.prgrms.mukvengers.domain.user.service.UserService;
